improve expand / collapse behavior simplify DOM structure all tests pass
